new structure for pages old files kept as reference, old folder should be deleted as soon as possible
