+++ pivotal_tracker/ui/features/project.feature
@@ -2,14 +2,13 @@ Feature: Create projects
   Background: preconditions
     Given I login the app as owner
 
-  @gui @clean_projects
-  Scenario: Verify that I can create a new Project
-     When I click on Create Project button
-      And I fill the form with data
-        | key           | value                                |
-        | project_name  | (prefix)_project_(current_date_time) |
-        | account       | (prefix)_account                     |
-        | private       | True                                 |
-        | public        | False                                |
-      And I click on Create button
-     Then I click on More button
+#  Scenario: Verify that new projects are created correctly
+#    When I create a project with characteristics
+#        | key           | value                                |
+#        | project_name  | (prefix)_project_(current_date_time) |
+#        | account       | (prefix)_account                     |
+#        | privacy       | public                               |
+#    Then I verify project name is displayed in header
+#    And I verify that project settings were created according to characteristics
+#    And I verify that project name is on dashboard
+#    And I verify that project is on projects menu
