+++ pivotal_tracker/ui/pages/project_main.py
@@ -1,0 +1,3 @@
+class ProjectMain:
+    def __init__(self, driver):
+        self.driver = driver
