+++ pivotal_tracker/ui/pages/tabs/dashboard_tabs.py
@@ -1,0 +1,3 @@
+class DashboardTabs(object):
+    PROJECTS = "Projects"
+    WORKSPACES = "Workspaces"
