+++ pivotal_tracker/ui/pages/tabs/user_main_tabs.py
@@ -1,0 +1,4 @@
+class UserMainTabs(object):
+    DASHBOARD = "Dashboard"
+    PROJECT_MAIN = "ProjectMain"
+    ALL_PROJECTS = "AllProjects"
