+++ pivotal_tracker/ui/pages/user_header.py
@@ -1,0 +1,3 @@
+class UserHeader:
+    def __init__(self):
+        pass
